README.md
A Selenium Data-Driven Project is an approach in which test data is separated from the test script, allowing for more flexibility and easier maintenance. The idea is to store test data externally, such as in a spreadsheet or database, and dynamically feed this data into the Selenium test script during execution. This approach is particularly useful when you need to execute the same test scenario with multiple sets of data.